ICU-22991 Simplified Grego code
Use timeToFields instead of dayToFields

+++ icu4c/source/i18n/gregoimp.h
@@ -110,7 +110,7 @@ class ClockMath {
      * Calling with a divisor <= 0 is disallowed.
      */
     static double floorDivide(double dividend, double divisor,
-                              double* remainder);
+                              int32_t* remainder);
 };
 
 // Useful millisecond constants
